requirements/base.txt
coreapi==2.3.0
Django==1.11.1
djangorestframework==3.6.3
